move Adapter processes to processes/adapter folder
