+++ src/component/processes/adapter/append.ts
@@ -1,6 +1,6 @@
-import { Scroller } from '../scroller';
-import { Item } from '../classes/item';
-import { Process, ProcessStatus } from '../interfaces/index';
+import { Scroller } from '../../scroller';
+import { Item } from '../../classes/item';
+import { Process, ProcessStatus } from '../../interfaces/index';
 
 export default class Append {
