+++ src/component/processes/adapter/check.ts
@@ -1,5 +1,5 @@
-import { Scroller } from '../scroller';
-import { Process, ProcessStatus } from '../interfaces/index';
+import { Scroller } from '../../scroller';
+import { Process, ProcessStatus } from '../../interfaces/index';
 
 export default class Check {
