+++ src/component/processes/adapter/reload.ts
@@ -1,6 +1,6 @@
-import { Scroller } from '../scroller';
-import { ADAPTER_METHODS } from '../inputs/index';
-import { Process, ProcessStatus } from '../interfaces/index';
+import { Scroller } from '../../scroller';
+import { ADAPTER_METHODS } from '../../inputs/index';
+import { Process, ProcessStatus } from '../../interfaces/index';
 
 export default class Reload {
